+++ nengo_loihi/hardware/nxsdk_shim.py
@@ -24,13 +24,7 @@ except ImportError:
         raise exception
 
 
-if HAS_NXSDK and nxsdk_version < LooseVersion("0.8.0"):  # pragma: no cover
-    import nxsdk.arch.n2a.compiler.microcodegen.interface as microcodegen_uci
-    from nxsdk.arch.n2a.compiler.tracecfggen.tracecfggen import TraceCfgGen
-    from nxsdk.arch.n2a.graph.graph import N2Board
-    from nxsdk.arch.n2a.graph.inputgen import BasicSpikeGenerator
-    from nxsdk.arch.n2a.graph.probes import N2SpikeProbe
-elif HAS_NXSDK:
+if HAS_NXSDK:
     import nxsdk.compiler.microcodegen.interface as microcodegen_uci
     from nxsdk.compiler.tracecfggen.tracecfggen import TraceCfgGen
     from nxsdk.graph.nxboard import N2Board